+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
 - Updated nginx configuration.
+- Made Memo Label optional for Fjernpost.
 
 ## [0.2.0]