+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated nginx configuration.
 - Made Memo Label optional for Fjernpost.
-- Made shipment order of letters stable.
+- Made shipment order of letters stable to ensure proper caching.
 
 ## [0.2.0]