+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Option to add attached files to Digital Shipments.
+- API endpoint to get attached files from a shipment.
+
 ### Changed
 
 - Updated nginx configuration.
 - Made Memo Label optional for Fjernpost.
+- Made shipment order of letters stable to ensure proper caching.
 
 ## [0.2.0]